pyproject.toml
[tool.poetry]
name = "ihs"
version = "0.1.0"
description = ""
authors = ["Brock Friedrich <brock@driftwoodenergy.com>"]

[tool.poetry.dependencies]
python = "^3.7"
pyyaml = "*"
pandas = "*"
sqlalchemy = "==1.2.18"
sentry-sdk = "*"
oauthlib = "*"
click = "*"
psycopg2-binary = "^2.8"
flask = "^1.1"
Flask-SQLAlchemy = "^2.4"
Flask-PyMongo = "*"
flask-debugtoolbar = "*"
flask-migrate = "*"
Flask-RESTful = "*"
python-dotenv = "^0.10.3"
attrdict = "^2.0"
requests_mock = "^1.7"
tomlkit = "^0.5.8"
logutils = "^0.3.5"
pyparsing = "^2.4"
attrs = "^19.3"
datadog = "^0.31.0"
celery = {version = "^4.3.0", extras = ["sqs"]}
pymongo = "^3.9.0"
zeep = "^3.4.0"
lxml = "^4.4.1"
xmltodict = "^0.12.0"


[tool.poetry.dev-dependencies]
black = { version = "*", allows-prereleases = true }
mypy = "*"
pylint = "*"
ipython = "*"
matplotlib = "*"
pytest = "^5.2"
requests-mock = "^1.7"
pytest-cov = "^2.8"
coverage = "^4.5"
docker = "^4.1.0"
boto3 = "^1.10.6"
hypothesis = "^4.42.7"


[tool.poetry.scripts]
iwell = "ihs.manage:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"


